update py clients
