+++ scripts/tts/talk.py
@@ -6,10 +6,28 @@ import time
 import wave
 import json
 from pathlib import Path
+import json
 
 import riva.client
 from riva.client.argparse_utils import add_connection_argparse_parameters
 
+def read_file_to_dict(file_path):
+    result_dict = {}
+    with open(file_path, 'r') as file:
+        for line_number, line in enumerate(file, start=1):
+            line = line.strip()
+            if not line:
+                print(f"Warning: Empty line at line number {line_number}.")
+                continue
+            try:
+                key, value = line.split('  ', 1)  # Split by double space
+                result_dict[str(key.strip())] = str(value.strip())
+            except ValueError:
+                print(f"Warning: Malformed line at line number {line_number}: {line}")
+                continue
+    if not result_dict:
+        raise ValueError("Error: No valid entries found in the file.")
+    return result_dict
 
 def parse_args() -> argparse.Namespace:
     parser = argparse.ArgumentParser(
@@ -42,7 +60,7 @@ def parse_args() -> argparse.Namespace:
     parser.add_argument(
         "--sample-rate-hz", type=int, default=44100, help="Number of audio frames per second in synthesized audio."
     )
-    parser.add_argument("--user-dictionary", type=str, help="Path to user dictionary file")
+    parser.add_argument("--user-dictionary", type=str, help="A file path to a user dictionary with key-value pairs separated by double spaces.")
     parser.add_argument(
         "--stream",
         action="store_true",
@@ -109,13 +127,16 @@ def main() -> None:
             out_f.setsampwidth(sampwidth)
             out_f.setframerate(args.sample_rate_hz)
 
+        if args.user_dictionary is not None:
+            user_dictionary_input = read_file_to_dict(args.user_dictionary)
+
         print("Generating audio for request...")
         start = time.time()
         if args.stream:
             responses = service.synthesize_online(
                 args.text, args.voice, args.language_code, sample_rate_hz=args.sample_rate_hz,
                 audio_prompt_file=args.audio_prompt_file, quality=20 if args.quality is None else args.quality,
-                user_dictionary=args.user_dictionary
+                user_dictionary=user_dictionary_input
             )
             first = True
             for resp in responses:
@@ -131,7 +152,7 @@ def main() -> None:
             resp = service.synthesize(
                 args.text, args.voice, args.language_code, sample_rate_hz=args.sample_rate_hz,
                 audio_prompt_file=args.audio_prompt_file, quality=20 if args.quality is None else args.quality,
-                user_dictionary=args.user_dictionary
+                user_dictionary=user_dictionary_input
             )
             stop = time.time()
             print(f"Time spent: {(stop - start):.3f}s")
